log gtags size

+++ scripts/macOnly/tutorialConfigUpdater.sh
@@ -45,22 +45,28 @@ prettyPrint "Copying scripts to custom Installer Repo $ZPWR_DIR"
 cp "$HOME/.gitignore_global" "$ZPWR_DIR_INSTALL"
 #cp -R "$HOME/.vim/Ultisnips" "$ZPWR_DIR_INSTALL"
 
-type=ctags
-
-prettyPrint "Regen GNU gtags to $HOME/GTAGS with $type parser"
-(
+if exists gtags; then
+    prettyPrint "Regen GNU gtags to $HOME/GTAGS with $type parser"
+    (
     builtin cd "$HOME"
     command rm GPATH GRTAGS GTAGS 2>/dev/null
     for file in \
         "$ZPWR_INSTALL/"* \
         "$ZPWR/"* \
         "$ZPWR_SCRIPTS"/* \
-        "$ZPWR_SCRIPTS/macOnly/"*; do
+        "$ZPWR_SCRIPTS_MAC/"*; do
         if [[ -f "$file" ]]; then
             echo "$file"
         fi
     done | gtags --accept-dotfiles --gtagslabel=$type -f -
-)
+
+    printf "gtags size: "
+    global -x ".*" | wc -l
+
+    )
+else
+    echo "gtags does not exist" >&2
+fi
 
 prettyPrint "Copying gtag => $HOME/"{GTAGS,GPATH,GRTAGS}
 echo cp "$HOME/"{GTAGS,GPATH,GRTAGS} \
